packages/sentinel/pyproject.toml
[tool.poetry]
name = "Sentinel"
version = "0.1.0"
description = "Home security and alert system."
authors = ["Dixon Sean Low Yan Feng", "Keith Neo Kai Si"]

[tool.poetry.dependencies]
sentinel-core = { path = "../sentinel-core", develop = true }
sentinel-opencv = { path = "../plugins/sentinel-opencv", develop = true }
sentinel-desktop-notification-subscriber = { path = "../plugins/sentinel-desktop-notification-subscriber", develop = true }
sentinel-ultralytics = { path = "../plugins/sentinel-ultralytics", develop = true }
sentinel-mediapipe = { path = "../plugins/sentinel-mediapipe", develop = true }
python = "^3.12"
opencv-python = "^4.10.0.84"
aioreactive = "^0.20.0"
expression = "^5.3.0"

[tool.poetry.scripts]
sentinel = "sentinel.cli:entry"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
